feat: [박성빈] 공통 컴포넌트로 사용할 FlexBox 컴포넌트를 작성합니다 [ZFQF-11]
